+++ src/routes/auth/index.ts
@@ -12,7 +12,10 @@ const router = Router();
 router.get(
 	'/google',
 	passport.authenticate('google', {
-		scope: ['https://www.googleapis.com/auth/userinfo.profile', 'https://www.googleapis.com/auth/userinfo.email']
+		scope: [
+			'https://www.googleapis.com/auth/userinfo.profile',
+			'https://www.googleapis.com/auth/userinfo.email'
+		]
 	})
 );